% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conprod_12.R
\docType{data}
\name{conprod_12}
\alias{conprod_12}
\title{Concentrate Production (kt)}
\format{
A dataframe with 26 observations and 2 variables:
\describe{
\item{Ni}{Nickel concentrate in kt}
\item{Cu}{Copper concentrate in kt}
}
}
\source{
Mine Plan
}
\usage{
data(conprod_12)
}
\description{
This dataset contains the annual Nickel (Ni) and Copper (Cu) concentrate production rates in kt for the 12mtpa mine plan.
}
\keyword{datasets}
