% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{Fuel used over life of mine}
\format{
A dataframe with 26 observations and 8 variables:
\describe{
\item{year}{year of mine operation}
\item{area}{area of mining operation, one of mine = "drilling and blasting", p_load = "primary loading", p_haul = "primary hauling", a_haul = "ancillary hauling", road = "road trucks", rail = "rail transport" and power = "power station".}
\item{diesel}{diesel annual usage in kL}
}
}
\source{
Engie Data Book Rev 2
}
\usage{
data(fuel)
}
\description{
This dataset contains the annual usage of diesel fuel, measured in kL, for different areas of operation.
}
\keyword{datasets}
