% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{Observations from West Musgrave LiDAR.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10810 rows and 105 columns.
}
\usage{
lidar
}
\description{
A dataset containing observations recorded from the on-site LiDAR every minute.
}
\keyword{datasets}
