% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metmast.R
\docType{data}
\name{metmast}
\alias{metmast}
\title{Observations from West Musgrave Met Mast.}
\format{
A data frame with 201232 rows and 37 variables:
\describe{
  \item{datetime}{timestamp of the observation}
  ...
}
}
\usage{
metmast
}
\description{
A dataset containing observations of meteorlogical variables every 10 minutes since May 2018.
}
\keyword{datasets}
