% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.R
\docType{data}
\name{plant}
\alias{plant}
\title{Processing Plant}
\format{
A dataframe with 52560 observations and 4 variables:
\describe{
\item{rop_12mtpa}{rest of plant load in MW at 12mtpa}
\item{rop_16mtpa}{rest of plant load in MW at 16mtpa}
\item{thru_12mtpa}{throughput in t at 12mtpa}
\item{thru_16mtpa}{throughput in t at 16mtpa}
}
}
\source{
Gres load list a base load profile build up
}
\usage{
data(plant)
}
\description{
This dataset includes various parameters for wind turbines under consideration at West Musgrave.
}
\keyword{datasets}
