% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV_model.R
\name{PV_model}
\alias{PV_model}
\title{PV model}
\usage{
PV_model(Pmax, Sr, Te, eff, stc_rad = 1000, stc_temp = 25)
}
\arguments{
\item{Pmax}{rated capacity.}

\item{Sr}{solar radiation.}

\item{Te}{air temperature in degrees celsius.}

\item{eff}{conversion efficiency.}
}
\value{
the power output of a PV module.
}
\description{
Calculates the power output of a PV module from solar radiation and air temperature. You can also change the STCs in order to alter the capacity factors.
}
\examples{
PV_model(1,800,35,0.8)
}
