% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sg.R
\docType{data}
\name{Sg}
\alias{Sg}
\title{Solar Generation}
\format{
A dataframe with 52,560 observations and 9 variables.
}
\source{
Original from OEMs: Siemens (hourly), Aurecon (10-min) and a rough model developed by OZ.
}
\usage{
data(Sg)
}
\description{
This dataset contains power output (MW) of various Solar Panels for every 10-minute increment in the 2018-2019 financial year. Original data was supplied by the OEMs who used met mast data, and power generation was then further subjected to losses around 8%.
}
\keyword{datasets}
