% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Te.R
\docType{data}
\name{Te}
\alias{Te}
\title{WM Air Temperature}
\format{
A dataframe with 153568 observations and 3 variables:
\describe{
\item{dtime}{datetime}
\item{AvgTe 97}{average air temperature at 97m.}
\item{AvgTe 5}{average air temperature at 5m.}
}
}
\source{
WM WindServer
}
\usage{
data(Te)
}
\description{
This dataset contains air temperature collected on-site at 10-minute intervals.
}
\keyword{datasets}
