% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wg.R
\docType{data}
\name{Wg}
\alias{Wg}
\title{Wind Generation}
\format{
A dataframe with 52,560 observations and 9 variables.
}
\source{
Original from OEMs: Vestas, Siemens, Acciona and a rough model developed by OZ.
}
\usage{
data(Wg)
}
\description{
This dataset contains power output (MW) of various WTGs for every 10-minute increment in the 2018-2019 financial year. Original data was supplied by the OEMs who used met mast data, and power generation was then further subjected to losses around 14%.
}
\keyword{datasets}
