% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTG_model.R
\name{wind_power_law}
\alias{wind_power_law}
\title{Power Law for Wind Speeds}
\usage{
wind_power_law(speed, height_obs, height_tar)
}
\arguments{
\item{speed}{wind speed (m/s)}

\item{height_obs}{height of observed wind speeds}

\item{height_tar}{height of target wind speeds}
}
\value{
wind speed (m/s) at target height
}
\description{
Estimate wind speeds at a target height based on wind speeds recorded at an observed height.
}
