% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windparams.R
\docType{data}
\name{windparams}
\alias{windparams}
\title{Parameters of Wind Turbines}
\format{
A dataframe with 27 observations and 3 variables:
\describe{
\item{name}{name of the turbine}
\item{param}{one of capex (capital costs), size (size of turbine in MW), and sweep (length of the blade in m)}
\item{value}{parameter value in the specified units}
}
}
\source{
Vestas, Siemens, and Acciona
}
\usage{
data(windparams)
}
\description{
This dataset includes various parameters for wind turbines under consideration at West Musgrave.
}
\keyword{datasets}
