% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ws.R
\docType{data}
\name{Ws}
\alias{Ws}
\title{WM Wind Speeds}
\format{
A dataframe with 148240 observations and 21 variables:
\describe{
\item{dtime}{datetime}
\item{AvgWs height direction}{average wind speed at height='height' and at direction='direction.}
\item{StdWs height direction}{standard deviation of wind speed at height='height' and at direction='direction.}
\item{MaxWs height direction}{maximum wind speed at height='height' and at direction='direction.}
\item{MinWs height direction}{Minimum wind speed at height='height' and at direction='direction.}
}
}
\source{
On-site Met Mast
}
\usage{
data(Ws)
}
\description{
This dataset contains wind speeds collected on-site at 10-minute intervals.
}
\keyword{datasets}
