% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{Observations from West Musgrave LiDAR.}
\format{
A data frame with 105 variables
\describe{
  \item{datetime}{timestamp of the observation}
  ...
}
}
\usage{
lidar
}
\description{
A dataset containing observations recorded from the on-site LiDAR every minute.
}
\keyword{datasets}
